// files/src/react/services/api-processor.enum.ts
export enum GeneralLedgerPayoutsStatus {
  PAYMENT_FILE_PROCESSING = 'PAYMENT_FILE_PROCESSING',
  PAYMENT_FILE_SUBMITTED = 'PAYMENT_FILE_SUBMITTED',
  ACK_SUCCEEDED = 'ACK_SUCCEEDED',
  ACK_PARTIAL_REJECTED = 'ACK_PARTIAL_REJECTED',
  ACK_REJECTED = 'ACK_REJECTED',
  ACK_INVALID = 'ACK_INVALID',
  PARTIAL_COMPLETED = 'PARTIAL_COMPLETED',
  COMPLETED = 'COMPLETED',
}

export enum GeneralLedgerPayoutTransactionStatus {
  PROCESSING = 'PROCESSING',
  PENDING_FOR_BANK = 'PENDING_FOR_BANK',
  SUCCEEDED = 'SUCCEEDED',
  FAILED = 'FAILED',
}

export enum PayoutBatchStatus {
  PAYMENT_FILE_PROCESSING = 'PAYMENT_FILE_PROCESSING',
  PAYMENT_FILE_SUBMITTED = 'PAYMENT_FILE_SUBMITTED',
  ACK_SUCCEEDED = 'ACK_SUCCEEDED',
  ACK_PARTIAL_REJECTED = 'ACK_PARTIAL_REJECTED',
  ACK_REJECTED = 'ACK_REJECTED',
  ACK_INVALID = 'ACK_INVALID',
  PARTIAL_COMPLETED = 'PARTIAL_COMPLETED',
  COMPLETED = 'COMPLETED',
}

export enum PayoutStatus {
  PROCESSING = 'PROCESSING',
  PENDING_FOR_BANK = 'PENDING_FOR_BANK',
  SUCCEEDED = 'SUCCEEDED',
  FAILED = 'FAILED',
}

export enum Mt940TransactionTypes {
  DEBIT = 'DEBIT',
  CREDIT = 'CREDIT',
}